# fix(batch): add first-match optimization for left semi join